chore: reorganize components file structure
