--- src/components/card.module.css
@@ -1,93 +1,0 @@
-.cardContainer {
-  position: relative;
-  display: inline-block;
-  cursor: pointer;
-}
-.cardContainer:hover .card {
-  transform: scale(1.02);
-}
-.cardContainer:hover .card::after {
-  opacity: 1;
-}
-.cardContainer:hover .cardBG {
-  opacity: 1;
-  visibility: visible;
-}
-.cardContainer:hover .icons {
-  opacity: 1;
-  transform: translateY(0);
-}
-.card {
-  background-color: rgb(54, 54, 54);
-  width: 154px;
-  height: 236px;
-  border-radius: 24px;
-  position: relative;
-  color: white;
-  margin-bottom: 10px;
-  transition: transform 0.1s ease-in-out, background-color 0.3s ease-in-out;
-  overflow: hidden;
-  background-size: cover;
-  background-position: center;
-}
-.card::after {
-  content: '';
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0, 0, 0, 0.6);
-  opacity: 0;
-  transition: opacity 0.3s ease-in-out;
-}
-.cardBG {
-  position: absolute;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  height: 100%;
-  width: 100%;
-  padding: 10px;
-  opacity: 0;
-  transition: opacity 0.3s ease-in-out, visibility 0.3s;
-  z-index: 1;
-  padding: 12px 8px;
-}
-.leftSide {
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-end;
-  align-items: flex-start;
-  height: 100%;
-}
-.tag {
-  position: absolute;
-  left: -6px;
-  top: 10px;
-  background-color: red;
-  border-radius: 6px;
-  padding: 2px 8px;
-  font-size: 12px;
-  z-index: 1;
-}
-.text {
-  display: flex;
-  flex-direction: column;
-  gap: 6px;
-  font-size: 14px;
-}
-.desc {
-  font-size: 10px;
-}
-.rightSide {
-  position: relative;
-}
-.name {
-  font-size: 12px;
-  padding-bottom: 2px;
-}
-.status {
-  font-size: 10px;
-}
